Fix example with bad practice checking existence of ref
React purposely passes null to ref functions when a component is unmounted so that you set whatever was referencing it to null, to ensure proper garbage collection. It's a bad idea to check the existence of the parameter before setting whatever you were gonna set -- you're *intended* to set it to null.